tmux: remove old mouse options
Although I could leave them in with -q to suppress the warning, and
it'd work fine in tmux 2.1+, the 'set -g -q mouse on' would break the
functionality in older tmux versions in which -q doesn't suppress
warnings. So mouse support on tmux on Ubuntu 14.04 wouldn't work at all
in that situation, and thus there's no point keeping it in.