Use colorscheme wombat at all times
The wombat-term.vim colorscheme had a bunch of if statements in there
which caused it to act weird when TERM=xterm-256color (which is the
current default for GNOME Terminal), and I really don't want to deal
with modifying the colorscheme to deal with that. I'll let CSApprox deal
with it.