Force screen off only when not locking immediately
This gives the user a heads-up that the screen is about to lock, giving
them some time to move the mouse. However, when wanting to lock
immediately, we should just set the timeouts and lock; turning the
screen off and then setting the timeouts (or vice versa) seems to cause
the screen to turn back on immediately.