Autostart ssh-agent if not already running
This makes ~/.ssh/ssh_auth_sock a symlink to $SSH_AUTH_SOCK, and starts
and ssh-agent if $SSH_AUTH_SOCK is not set. This means that on regular
machines everything works the same as before (albeit using a new
symlink), but on machines I just ssh into, I'll always have an ssh-agent
running.